Updated SQL Dilects and Inspections
