+++ .idea/sqldialects.xml
@@ -2,5 +2,7 @@
 <project version="4">
   <component name="SqlDialectMappings">
     <file url="file://$PROJECT_DIR$/server/src/main/java/sql/MySQL.java" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/server/src/test/java/sql/MySqlTest.java" dialect="GenericSQL" />
+    <file url="PROJECT" dialect="MariaDB" />
   </component>
 </project>
